Move argparse and printing version info to workflow_helpers
Now, there is only one general argparse function that can be accessed by all scripts.
