+++ workflow/magi_workflow_gene_to_reaction.py
@@ -302,16 +302,20 @@ def main():
     main_start = time.time() # overall program timer
     
     # Parse arguments and prepare for gene to reaction workflow
-    args = parse_arguments()
-    output_dir, intermediate_files_dir = mg.make_output_dirs(output_dir=args.output, fasta_file=args.fasta, intermediate_files = args.intermediate_files)
+    magi_parameters = magi.general_magi_preparation()
     
     #Run gene to reaction workflow
-    gene_to_reaction_top_path, gene_database_path = workflow(args.fasta, intermediate_files_dir, args.cpu_count,
-             args.annotations, args.blast_filter)
+    workflow(fasta_file=magi_parameters["fasta"], 
+             intermediate_files_dir=magi_parameters["intermediate_files_dir"], 
+             cpu_count=magi_parameters["cpu_count"],
+             annotations=magi_parameters["annotations"], 
+             blast_filter=magi_parameters["blast_filter"])
     
     #Format output if this is the last step of the workflow
-    if args.gene_to_reaction_only:
-        format_output(gene_to_reaction_top_path, output_dir, main_start)
+    if magi_parameters["gene_to_reaction_only"]:
+        g2r_file = os.path.join(magi_parameters["intermediate_files_dir"], 
+                                            'gene_to_reaction.pkl')
+        format_output(g2r_file, magi_parameters["output_dir"], main_start)
 
 if __name__ == "__main__":
     main()
